Adds model probability method to the package and makes corrections to the implied probability calculator.
